+++ j2k/src/org/jetbrains/kotlin/j2k/Utils.kt
@@ -82,6 +82,17 @@ fun PsiElement.getContainingMethod(): PsiMethod? {
     return null
 }
 
+fun PsiElement.getContainingClass(): PsiClass? {
+    var context = context
+    while (context != null) {
+        val _context = context
+        if (_context is PsiClass) return _context
+        if (_context is PsiMember) return _context.containingClass
+        context = _context.context
+    }
+    return null
+}
+
 fun PsiElement.getContainingConstructor(): PsiMethod? {
     val method = getContainingMethod()
     return if (method?.isConstructor == true) method else null