+++ j2k/testData/fileOrElement/protected/onlyProtectedConstructor.java
@@ -1,0 +1,13 @@
+package test;
+
+public class BaseProtectedConstructor {
+    protected BaseSamePackage() {
+
+    }
+}
+
+class DerivedSamePackage extends BaseProtectedConstructor {
+    DerivedSamePackage() {
+        super();
+    }
+}