+++ j2k/testData/multiFile/ProtectedVisibility/InheritorOtherPackage1.java
@@ -1,0 +1,13 @@
+package test;
+
+public class BaseOtherPackage {
+    protected BaseOtherPackage() {
+
+    }
+
+    protected void foo() {
+
+    }
+
+    protected int i = 1;
+}