+++ j2k/testData/multiFile/ProtectedVisibility/InheritorOtherPackage2.java
@@ -1,0 +1,11 @@
+package test2;
+
+import test.*;
+
+public class DerivedOtherPackage extends BaseOtherPackage {
+    public void usage1() {
+        BaseOtherPackage base = new BaseOtherPackage();
+        base.foo();
+        int i = base.i;
+    }
+}