+++ Changelog.md
@@ -6,3 +6,6 @@
 
 ### JS
 - Safe calls (`x?.let { it }`) are now inlined
+
+### Tools. J2K
+- Protected members used outside of inheritors are converted as public